Use IC as the platform type code
Platform Type is a code that determines which IDE the platform represents (e.g., IDEA, Android Studio, PyCharm, etc).  The code 'MP' is not recognized by the IntelliJ Platform Gradle Plugin, so it gives errors when building the plugin against our custom platform. 

The code for IntellliJ IDEA is `IC`, so let's use that one.